refactor(file-client): move shared constants into separate file
