+++ lib/file-client/file-constants.ts
@@ -1,0 +1,8 @@
+/** Directory where repositories are cloned to */
+export const CACHE_LOCATION = './.cache-eslint-remote-tester';
+
+/** Directory where results are generated in to */
+export const RESULTS_LOCATION = './eslint-remote-tester-results';
+
+/** Base URL for repository cloning */
+export const URL = 'https://github.com';
