+++ lib/file-client/index.ts
@@ -2,9 +2,9 @@ export { getFiles, SourceFile } from './file-client';
 export {
     writeResults,
     prepareResultsDirectory,
-    RESULTS_LOCATION,
     RESULT_TEMPLATE,
 } from './results-writer';
-export { CACHE_LOCATION, removeCachedRepository } from './repository-client';
+export { CACHE_LOCATION, RESULTS_LOCATION } from './file-constants';
+export { removeCachedRepository } from './repository-client';
 export { default as ResultsStore } from './results-store';
 export { RESULTS_TEMPLATE_CI_BASE } from './result-templates';
